+++ conformance/fhir-ig-davinci-pdex-formulary/src/test/java/com/ibm/fhir/ig/davinci/pdex/formulary/test/ConstraintGeneratorTest.java
@@ -10,7 +10,7 @@ import static com.ibm.fhir.path.util.FHIRPathUtil.compile;
 
 import org.testng.annotations.Test;
 
-import com.ibm.fhir.ig.davinci.pdex.formulary.FormularyResourceProvider;
+import com.ibm.fhir.ig.davinci.pdex.formulary.Formulary101ResourceProvider;
 import com.ibm.fhir.model.annotation.Constraint;
 import com.ibm.fhir.model.resource.StructureDefinition;
 import com.ibm.fhir.model.type.Extension;
@@ -21,11 +21,11 @@ import com.ibm.fhir.registry.spi.FHIRRegistryResourceProvider;
 
 public class ConstraintGeneratorTest {
     @Test
-    public static void testConstraintGenerator() throws Exception {
-        FHIRRegistryResourceProvider provider = new FormularyResourceProvider();
+    public static void test101ConstraintGenerator() throws Exception {
+        FHIRRegistryResourceProvider provider = new Formulary101ResourceProvider();
         for (FHIRRegistryResource registryResource : provider.getRegistryResources()) {
             if (StructureDefinition.class.equals(registryResource.getResourceType())) {
-                String url = registryResource.getUrl();
+                String url = registryResource.getUrl() + "|" + registryResource.getVersion();
                 System.out.println(url);
                 Class<?> type = ModelSupport.isResourceType(registryResource.getType()) ? ModelSupport.getResourceType(registryResource.getType()) : Extension.class;
                 for (Constraint constraint : ProfileSupport.getConstraints(url, type)) {
@@ -37,5 +37,36 @@ public class ConstraintGeneratorTest {
                 }
             }
         }
+    }
+
+    @Test
+    public static void test110ConstraintGenerator() throws Exception {
+        FHIRRegistryResourceProvider provider = new Formulary101ResourceProvider();
+        for (FHIRRegistryResource registryResource : provider.getRegistryResources()) {
+            if (StructureDefinition.class.equals(registryResource.getResourceType())) {
+                String url = registryResource.getUrl() + "|" + registryResource.getVersion();
+                System.out.println(url);
+                Class<?> type = ModelSupport.isResourceType(registryResource.getType()) ? ModelSupport.getResourceType(registryResource.getType()) : Extension.class;
+                for (Constraint constraint : ProfileSupport.getConstraints(url, type)) {
+                    System.out.println("    " + constraint);
+                    if (!Constraint.LOCATION_BASE.equals(constraint.location())) {
+                        compile(constraint.location());
+                    }
+                    compile(constraint.expression());
+                }
+            }
+        }
+    }
+
+    public static void main(String[] args) {
+        String url = "http://hl7.org/fhir/us/davinci-drug-formulary/StructureDefinition/usdf-DrugTierDefinition-extension|1.0.1";
+        for (Constraint constraint : ProfileSupport.getConstraints(url, Extension.class)) {
+            System.out.println("    " + constraint);
+            if (!Constraint.LOCATION_BASE.equals(constraint.location())) {
+                compile(constraint.location());
+            }
+            compile(constraint.expression());
+        }
+
     }
 }